Create and use email line item dates mixin
