+++ app/components/email-line-item/dates.js
@@ -1,62 +1,8 @@
 import Component from '@ember/component';
-import { computed } from '@ember/object';
-import UTCDatesMixin from 'fortnight/mixins/utc-dates-mixin';
-import dayjs from 'fortnight/dayjs';
+import LineItemDatesMixin from 'fortnight/mixins/line-item-dates-mixin';
 
-const { isArray } = Array;
-
-export default Component.extend(UTCDatesMixin, {
+export default Component.extend(LineItemDatesMixin, {
   classNames: ['form-group'],
-
-  start: null,
-  end: null,
-  days: null,
-  type: null,
-
-  format: 'ddd MMM Do, YYYY',
-
-  showDateRange: computed('type', function() {
-    return this.get('type') === 'range';
-  }),
-
-  _days: computed('days.[]', function() {
-    return this.get('days').map(timestamp => this.utcDateFromTimestamp(timestamp));
-  }),
-
-  _range: computed('start', 'end', function() {
-    const start = this.get('start');
-    const end = this.get('end');
-    return {
-      start: start ? this.utcDateFromTimestamp(start) : null,
-      end: end ? this.utcDateFromTimestamp(end) : null,
-    }
-  }),
-
-  daysSorted: computed('_days.[]', function() {
-    return this.get('_days').sort((a, b) => {
-      if (a.valueOf() > b.valueOf()) return 1;
-      if (a.valueOf() < b.valueOf()) return -1;
-      return 0;
-    });
-  }),
-
-  formattedDays: computed('daysSorted.[]', function() {
-    return this.get('daysSorted').map(d => dayjs(d).format(this.format));
-  }),
-
-  formattedRange: computed('_range.{start,end}', function() {
-    const start = this.get('_range.start');
-    const end = this.get('_range.end');
-    return {
-      start: start ? dayjs(start).format(this.format) : null,
-      end: end ? dayjs(end).format(this.format) : null,
-    }
-  }),
-
-  isRangeSet: computed('start', 'end', function() {
-    if (this.get('start') && this.get('end')) return true;
-    return false;
-  }),
 
   init() {
     this._super(...arguments);
@@ -64,10 +10,6 @@ export default Component.extend(UTCDatesMixin, {
       { key: 'range', label: 'By Range' },
       { key: 'days', label: 'By Days' },
     ]);
-    const days = this.get('days');
-    this.set('days', isArray(days) ? days : []);
-    const type = this.get('type');
-    this.set('type', type || 'days');
   },
 
   actions: {
